+++ src/helpers/build-field-definition.ts
@@ -11,21 +11,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { buildTypeMetadata } from "./build-type-metadata";
+import { buildTypeMetadata, TypeMetadata } from "./build-type-metadata";
 import {
   FieldDefinitionNode,
   GraphQLSchema,
   InterfaceTypeDefinitionNode,
   Kind,
   ObjectTypeDefinitionNode,
+  isInterfaceType,
 } from "graphql";
 import { CodegenConfigWithDefaults } from "./build-config-with-defaults";
+import { isExternalField } from "./is-external-field";
+import { indent } from "@graphql-codegen/visitor-plugin-common";
+import { buildAnnotations } from "./build-annotations";
 
 export function buildFieldDefinition(
   node: ObjectTypeDefinitionNode | InterfaceTypeDefinitionNode,
   fieldNode: FieldDefinitionNode,
   schema: GraphQLSchema,
   config: CodegenConfigWithDefaults,
+  typeMetadata: TypeMetadata,
 ) {
   const modifier = "val";
   const existingFieldArguments = fieldNode.arguments?.map((arg) => {
@@ -37,7 +42,38 @@ export function buildFieldDefinition(
   const fieldArguments = allFieldArguments?.length
     ? `(${allFieldArguments?.join(", ")})`
     : "";
-  return `${modifier} ${fieldNode.name.value}${fieldArguments}`;
+  const fieldDefinition = `${modifier} ${fieldNode.name.value}${fieldArguments}`;
+  const annotations = buildAnnotations({
+    config,
+    definitionNode: fieldNode,
+    typeMetadata,
+  });
+  if (node.kind === Kind.INTERFACE_TYPE_DEFINITION) {
+    const fieldText = indent(
+      `${fieldDefinition}: ${typeMetadata.typeName}${
+        typeMetadata.isNullable ? "?" : ""
+      }`,
+      2,
+    );
+    return `${annotations}${fieldText}`;
+  }
+  // const completableFutureDefinition = `java.util.concurrent.CompletableFuture<${typeMetadata.typeName}${typeMetadata.isNullable ? "?" : ""}>`;
+  const defaultValue = typeMetadata.defaultValue;
+  const defaultDefinition = `${typeMetadata.typeName}${isExternalField(fieldNode) ? (typeMetadata.isNullable ? "?" : "") : defaultValue}`;
+  const shouldOverrideField = node.interfaces?.some((interfaceNode) => {
+    const typeNode = schema.getType(interfaceNode.name.value);
+    return (
+      isInterfaceType(typeNode) &&
+      typeNode.astNode?.fields?.some(
+        (field) => field.name.value === fieldNode.name.value,
+      )
+    );
+  });
+  const field = indent(
+    `${shouldOverrideField ? "override " : ""}${fieldDefinition}: ${defaultDefinition}`,
+    2,
+  );
+  return `${annotations}${field}`;
 }
 
 export function buildFunctionFieldDefinition(
@@ -45,9 +81,20 @@ export function buildFunctionFieldDefinition(
   fieldNode: FieldDefinitionNode,
   schema: GraphQLSchema,
   config: CodegenConfigWithDefaults,
+  typeMetadata: TypeMetadata,
   completableFuture?: boolean,
-  shouldOverrideField?: boolean,
 ) {
+  const shouldOverrideField =
+    !completableFuture &&
+    node.interfaces?.some((interfaceNode) => {
+      const typeNode = schema.getType(interfaceNode.name.value);
+      return (
+        isInterfaceType(typeNode) &&
+        typeNode.astNode?.fields?.some(
+          (field) => field.name.value === fieldNode.name.value,
+        )
+      );
+    });
   const modifier =
     completableFuture ||
     node.kind === Kind.INTERFACE_TYPE_DEFINITION ||
@@ -65,5 +112,28 @@ export function buildFunctionFieldDefinition(
   const fieldArguments = allFieldArguments?.length
     ? `(${allFieldArguments?.join(", ")})`
     : "()";
-  return `${modifier} ${fieldNode.name.value}${fieldArguments}`;
+  const fieldDefinition = `${modifier} ${fieldNode.name.value}${fieldArguments}`;
+  const annotations = buildAnnotations({
+    config,
+    definitionNode: fieldNode,
+    typeMetadata,
+  });
+  if (node.kind === Kind.INTERFACE_TYPE_DEFINITION) {
+    const fieldText = indent(
+      `${fieldDefinition}: ${typeMetadata.typeName}${
+        typeMetadata.isNullable ? "?" : ""
+      }`,
+      2,
+    );
+    return `${annotations}${fieldText}`;
+  }
+
+  const completableFutureDefinition = `java.util.concurrent.CompletableFuture<${typeMetadata.typeName}${typeMetadata.isNullable ? "?" : ""}>`;
+  const defaultValue = `${typeMetadata.isNullable ? "?" : ""} = throw NotImplementedError("${node.name.value}.${fieldNode.name.value} must be implemented.")`;
+  const defaultDefinition = `${typeMetadata.typeName}${isExternalField(fieldNode) ? (typeMetadata.isNullable ? "?" : "") : defaultValue}`;
+  const field = indent(
+    `${shouldOverrideField ? "override " : ""}${fieldDefinition}: ${completableFuture ? completableFutureDefinition : defaultDefinition}`,
+    2,
+  );
+  return `${annotations}${field}`;
 }
